// packages/warps/src/types/warp.ts
export type WarpInitConfig = {
  env: WarpChainEnv
  clientUrl?: string
  currentUrl?: string
  vars?: Record<string, string | number>
  user?: {
    wallet?: string
  }
  schema?: {
    warp?: string
    brand?: string
  }
  cache?: {
    ttl?: number
    type?: CacheType
  }
  registry?: {
    contract?: string
  }
  index?: {
    url?: string
    apiKey?: string
    searchParamName?: string
  }
}

export type WarpCacheConfig = {
  ttl?: number
}

export type WarpChainInfo = {
  name: WarpChain
  displayName: string
  chainId: string
  blockTime: number
  addressHrp: string
  apiUrl: string
  explorerUrl: string
  nativeToken: string
}

export type WarpIdType = 'hash' | 'alias'

export type WarpVarPlaceholder = string

export type WarpResultName = string

export type WarpResulutionPath = string

export type WarpMessageName = string

export type Warp = {
  protocol: string
  name: string
  title: string
  description: string | null
  bot?: string
  preview?: string
  vars?: Record<WarpVarPlaceholder, string>
  actions: WarpAction[]
  next?: string
  results?: Record<WarpResultName, WarpResulutionPath>
  messages?: Record<WarpMessageName, string>
  meta?: WarpMeta
}

export type WarpMeta = {
  hash: string
  creator: string
  createdAt: string
}

export type WarpAction = WarpTransferAction | WarpContractAction | WarpQueryAction | WarpCollectAction | WarpLinkAction

export type WarpActionType = 'transfer' | 'contract' | 'query' | 'collect' | 'link'

export type WarpTransferAction = {
  type: WarpActionType
  chain?: WarpChain
  label: string
  description?: string | null
  address?: string
  data?: string
  value?: string
  transfers?: WarpContractActionTransfer[]
  inputs?: WarpActionInput[]
  next?: string
}

export type WarpContractAction = {
  type: WarpActionType
  chain?: WarpChain
  label: string
  description?: string | null
  address?: string
  func?: string | null
  args?: string[]
  value?: string
  gasLimit: number
  transfers?: WarpContractActionTransfer[]
  abi?: string
  inputs?: WarpActionInput[]
  next?: string
}

export type WarpContractActionTransfer = {
  token: string
  nonce?: number
  amount?: string
}

export type WarpQueryAction = {
  type: WarpActionType
  chain?: WarpChain
  label: string
  description?: string | null
  address?: string
  func?: string
  args?: string[]
  abi?: string
  inputs?: WarpActionInput[]
  next?: string
}

export type WarpCollectAction = {
  type: WarpActionType
  chain?: WarpChain
  label: string
  description?: string | null
  destination: {
    url: string
    method?: 'GET' | 'POST' | 'PUT' | 'DELETE'
    headers?: Record<string, string>
  }
  inputs?: WarpActionInput[]
  next?: string
}

export type WarpLinkAction = {
  type: WarpActionType
  chain?: WarpChain
  label: string
  description?: string | null
  url: string
  inputs?: WarpActionInput[]
}